# docs/swagger.yaml
definitions:
...
  models.TaskEvent:
    properties:
      actor_id:
        type: integer
      created_at:
        type: string
      event_id:
        type: string
      event_type:
        type: string
      id:
        type: integer
      payload:
        type: object
      project_id:
        type: integer
      task_id:
        type: integer
      task_version:
        type: integer
    type: object
...
  response.Resp:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.ProjectSyncResult:
    properties:
      events:
        items:
          $ref: '#/definitions/models.TaskEvent'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /projects/{id}/sync:
    get:
      consumes:
      - application/json
      description: Returns task event rows after a cursor for reconnect/backfill.
        The cursor is the task_events auto-increment id.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Last seen task event id
        in: query
        name: cursor
        type: integer
      - description: Page size, max 200
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sync events loaded
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/service.ProjectSyncResult'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/response.Resp'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/response.Resp'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: Sync project task events
      tags:
      - Realtime
...
swagger: "2.0"